organize raycasters files
