+++ genesis/sensors/raycaster/__init__.py
@@ -1,4 +1,11 @@
-from .base_pattern import RaycastPattern
-from .camera_pattern import DepthCameraPattern
-from .lidar_pattern import GridPattern, LidarPattern, SphericalPattern, SpinningLidarPattern
-from .raycaster import DepthCameraOptions, LidarOptions, RaycasterOptions
+from .depth_camera import DepthCameraOptions
+from .lidar import LidarOptions
+from .patterns import (
+    DepthCameraPattern,
+    GridPattern,
+    LidarPattern,
+    RaycastPattern,
+    SphericalPattern,
+    SpinningLidarPattern,
+)
+from .raycaster import RaycasterOptions
