+++ genesis/sensors/raycaster/patterns/base_pattern.py
