+++ genesis/sensors/raycaster/patterns/camera_patterns.py
