+++ genesis/sensors/raycaster/raycaster.py
@@ -11,6 +11,7 @@ import genesis as gs
 import genesis.engine.solvers.rigid.rigid_solver_decomp as rigid_solver_decomp
 import genesis.utils.array_class as array_class
 from genesis.engine.bvh import AABB, LBVH
+from genesis.sensors.sensor_manager import register_sensor
 from genesis.utils.geom import (
     ti_normalize,
     ti_transform_by_quat,
@@ -29,13 +30,7 @@ from ..base_sensor import (
     SensorOptions,
     SharedSensorMetadata,
 )
-from ..sensor_manager import register_sensor
-from .base_pattern import (
-    RaycastPattern,
-    RaycastPatternGenerator,
-    create_pattern_generator,
-)
-from .camera_pattern import DepthCameraPattern
+from .patterns import RaycastPattern, RaycastPatternGenerator, create_pattern_generator
 
 if TYPE_CHECKING:
     from genesis.utils.ring_buffer import TensorRingBuffer
@@ -58,7 +53,6 @@ def ray_triangle_intersection(ray_start, ray_dir, v0, v1, v2):
     """
     result = ti.math.vec4(0.0, 0.0, 0.0, 0.0)
 
-    # Compute edge vectors
     edge1 = v1 - v0
     edge2 = v2 - v0
 
@@ -69,7 +63,6 @@ def ray_triangle_intersection(ray_start, ray_dir, v0, v1, v2):
     # Check all conditions in sequence without early returns
     valid = True
 
-    # Declare all variables at the top to avoid scope issues
     t = 0.0
     u = 0.0
     v = 0.0
@@ -78,7 +71,7 @@ def ray_triangle_intersection(ray_start, ray_dir, v0, v1, v2):
     q = ti.math.vec3(0.0, 0.0, 0.0)
 
     # If determinant is near zero, ray lies in plane of triangle
-    if ti.abs(a) < 1e-8:
+    if ti.abs(a) < gs.EPS:
         valid = False
 
     if valid:
@@ -86,7 +79,6 @@ def ray_triangle_intersection(ray_start, ray_dir, v0, v1, v2):
         s = ray_start - v0
         u = f * s.dot(h)
 
-        # Check u parameter bounds
         if u < 0.0 or u > 1.0:
             valid = False
 
@@ -94,7 +86,6 @@ def ray_triangle_intersection(ray_start, ray_dir, v0, v1, v2):
         q = s.cross(edge1)
         v = f * ray_dir.dot(q)
 
-        # Check v parameter bounds
         if v < 0.0 or u + v > 1.0:
             valid = False
 
@@ -103,10 +94,9 @@ def ray_triangle_intersection(ray_start, ray_dir, v0, v1, v2):
         t = f * edge2.dot(q)
 
         # Ray intersection
-        if t <= gs.EPS:  # Invalid intersection
+        if t <= gs.EPS:
             valid = False
 
-    # Set result only if valid
     if valid:
         result = ti.math.vec4(t, u, v, 1.0)
 
@@ -556,123 +546,3 @@ class RaycasterSensor(RigidSensorMixin, Sensor):
             self.debug_objects[env_idx] = context.draw_debug_spheres(
                 points, radius=self._options.debug_sphere_radius, color=color
             )
-
-
-class LidarOptions(RaycasterOptions):
-    """
-    Lidar sensor that performs ray casting to get distance measurements and point clouds.
-
-    Parameters
-    ----------
-    entity_idx : int
-        The global entity index of the RigidEntity to which this sensor is attached.
-    link_idx_local : int, optional
-        The local index of the RigidLink of the RigidEntity to which this sensor is attached.
-    pos_offset : tuple[float, float, float], optional
-        The mounting offset position of the sensor in the world frame. Defaults to (0.0, 0.0, 0.0).
-    euler_offset : tuple[float, float, float], optional
-        The mounting offset quaternion of the sensor in the world frame. Defaults to (0.0, 0.0, 0.0).
-    pattern: RaycasterPattern
-        The raycasting pattern configuration for the sensor.
-    min_range : float, optional
-        The minimum sensing range in meters. Defaults to 0.0.
-    max_range : float, optional
-        The maximum sensing range in meters. Defaults to 20.0.
-    no_hit_value : float, optional
-        The value to return for no hit. Defaults to max_range if not specified.
-    return_world_frame : bool, optional
-        Whether to return points in the world frame. Defaults to False (local frame).
-    only_cast_fixed : bool, optional
-        Whether to only cast rays on fixed geoms. Defaults to False. This is a shared option, so the value of this
-        option for the **first** lidar sensor will be the behavior for **all** Lidar sensors.
-    resolution: float, optional
-        The resolution of the range measurement. Defaults to 1e-3.
-    noise: float, optional
-        The standard deviation of the noise added to the range measurement.
-    bias : float, optional
-        The bias added to the range measurement.
-    random_walk_std : float, optional
-        The standard deviation of the random walk added to the range measurement.
-    delay : float, optional
-        The delay in seconds before the sensor data is read.
-    jitter : float, optional
-        The time jitter standard deviation in seconds before the sensor data is read.
-    interpolate_for_delay : bool, optional
-        If True, the sensor data is interpolated between data points for delay + jitter.
-        Otherwise, the sensor data at the closest time step will be used. Default is False.
-    update_ground_truth_only : bool, optional
-        If True, the sensor will only update the ground truth cache, and not the measured cache.
-    """
-
-    pass
-
-
-@register_sensor(LidarOptions, RaycasterSharedMetadata)
-class LidarSensor(RaycasterSensor):
-    pass
-
-
-class DepthCameraOptions(RaycasterOptions):
-    """
-    Depth camera that uses ray casting to obtain depth images.
-
-    Parameters
-    ----------
-    entity_idx : int
-        The global entity index of the RigidEntity to which this sensor is attached.
-    link_idx_local : int, optional
-        The local index of the RigidLink of the RigidEntity to which this sensor is attached.
-    pos_offset : tuple[float, float, float], optional
-        The mounting offset position of the sensor in the world frame. Defaults to (0.0, 0.0, 0.0).
-    euler_offset : tuple[float, float, float], optional
-        The mounting offset quaternion of the sensor in the world frame. Defaults to (0.0, 0.0, 0.0).
-    pattern: RaycasterPattern
-        The raycasting pattern configuration for the sensor.
-    min_range : float, optional
-        The minimum sensing range in meters. Defaults to 0.0.
-    max_range : float, optional
-        The maximum sensing range in meters. Defaults to 20.0.
-    return_world_frame : bool, optional
-        Whether to return points in the world frame. Defaults to False (local frame).
-    only_cast_fixed : bool, optional
-        Whether to only cast rays on fixed geoms. Defaults to False. This is a shared option, so the value of this
-        option for the **first** lidar sensor will be the behavior for **all** Lidar sensors.
-    resolution: float, optional
-        The resolution of the range measurement. Defaults to 1e-3.
-    noise: float, optional
-        The standard deviation of the noise added to the range measurement.
-    bias : float, optional
-        The bias added to the range measurement.
-    random_walk_std : float, optional
-        The standard deviation of the random walk added to the range measurement.
-    delay : float, optional
-        The delay in seconds before the sensor data is read.
-    jitter : float, optional
-        The time jitter standard deviation in seconds before the sensor data is read.
-    interpolate_for_delay : bool, optional
-        If True, the sensor data is interpolated between data points for delay + jitter.
-        Otherwise, the sensor data at the closest time step will be used. Default is False.
-    update_ground_truth_only : bool, optional
-        If True, the sensor will only update the ground truth cache, and not the measured cache.
-    """
-
-    def validate(self, scene):
-        super().validate(scene)
-        if not isinstance(self.pattern, DepthCameraPattern):
-            gs.raise_exception("DepthCamera pattern must be an instance of DepthCameraPattern")
-
-
-@register_sensor(DepthCameraOptions, RaycasterSharedMetadata)
-class DepthCameraSensor(RaycasterSensor):
-    def read_image(self) -> torch.Tensor:
-        """
-        Read the depth image from the sensor.
-
-        This method uses the hit distances from the underlying RaycasterSensor.read() method and reshapes into image.
-
-        Returns
-        -------
-        torch.Tensor
-            The depth image with shape (height, width).
-        """
-        return self.read()["hit_ranges"].reshape(self._options.pattern.height, self._options.pattern.width)
